Update Enchant.js
Feel free to remove anything you don't agree with.  Just made a few additions and fixed a couple small bugs I found.  I have this bot running over on West and I'll probably continue to enhance it.  Let me know if there is anything in this check in that is wrong or unappealing either at bh, projectetal or d2jsp.  Thx. ~frickaline